Adopt the dst_version 24 tokens_info layout and skip_token_section
parser, and keep FT TX collect, RX derive, and apply_rx_tokens on
that handshake so extract stays compatible with the upstreamable
wire protocol.

+++ src/ucp/core/ucp_ep.h
@@ -517,8 +517,10 @@ typedef struct ucp_ep_recovery_arg {
     /* number of retries left before giving up */
     unsigned                retries_left;
     uint8_t                 state;
-    /* ADDR handshake generation for TX/RX token correlation */
-    uint64_t                request_id;
+    /* Generation of the LANES_ADDR exchange, pre-incremented by every request
+     * and echoed by the peer in its answers. Matched against ADDR RX tokens
+     * in ucp_ep_failover_apply_rx_tokens(). */
+    uint32_t                request_id;
     ucp_ep_recovery_probe_t probe[UCP_MAX_LANES];
 } ucp_ep_recovery_arg_t;
